+++ modules/com_engine/builtin/cpp/COM_dispBuiltin.cpp
@@ -35,7 +35,7 @@ Nelson::ComEngineGateway::COM_dispBuiltin(Evaluator* eval, int nLhs, const Array
     ArrayOfVector retval;
 #ifdef _MSC_VER
     nargoutcheck(nLhs, 0, 0);
-    nargincheck(argIn, 1, 1);
+    nargincheck(argIn, 1, 2);
     ArrayOf param1 = argIn[0];
     DispComHandleObject(eval, param1);
 #else