+++ modules/dynamic_link/builtin/cpp/libpointer_dispBuiltin.cpp
@@ -36,7 +36,7 @@ Nelson::DynamicLinkGateway::libpointer_dispBuiltin(
     Evaluator* eval, int nLhs, const ArrayOfVector& argIn)
 {
     ArrayOfVector retval;
-    nargincheck(argIn, 1, 1);
+    nargincheck(argIn, 1, 2);
     nargoutcheck(nLhs, 0, 0);
     ArrayOf param1 = argIn[0];
     if (param1.isHandle()) {